ENHANCEMENT: Added flicker free redraw to ImageEditor.plug-in.py
The image editor was flickering badly on Win7x64, Aero disabled
so I added flicker free redraws by:
1. Ignoring the EraseBackground event for the edit window
2. Buffering the draw to an AutoBufferedPaintDC (this is a nop on
   OSs that already have doublebuffered drawing)
3. Clearing the background in OnEditWindowPaint

Other option would be to only clear to background the parts of the
window that won't be covered by the bitmap, but that doesn't prevent
the grid xor drawing on top from causing flicker.

Tests:
Loaded several png and jpg files, verified there's no flicker and
that the background is cleared properly with and without scrollbars.